requirements.txt
# Система генерации форекс сигналов - Зависимости
# Основные библиотеки
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
aiohttp>=3.8.0

# Технический анализ
TA-Lib>=0.4.25

# JWT Authentication
PyJWT>=2.8.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# Environment
python-dotenv>=1.0.0

# Опциональные зависимости для расширенного функционала
# Раскомментируйте при необходимости:

matplotlib>=3.7.0  # Для графиков
# plotly>=5.15.0     # Для интерактивных графиков
# scipy>=1.11.0      # Для статистических расчетов
# scikit-learn>=1.3.0  # Для машинного обучения

# Утилиты
python-dateutil>=2.8.0
pytz>=2023.3
